+++ app/requests/detail/[id]/page.tsx
@@ -1,12 +1,12 @@
 import { getRequest, loadAllRequests } from '../../../../components/chain_utils';
 import DetailPanel from './panel';
 
-export async function generateStaticParams() {
-  const requests = await loadAllRequests();
-  return requests.map(request => ({
-      id: request.id,
-  }));
-}
+// export async function generateStaticParams() {
+//   const requests = await loadAllRequests();
+//   return requests.map(request => ({
+//       id: request.id,
+//   }));
+// }
 
 export default async function Page({ params }){
   const requestID = params.id;